Move launch feedback into cursors KCM
While doing some tree testing to evaluate the effectiveness
of the arrangement of System Settings' arrangement for actual
users and not just designers, I noticed that users consistently
expected launch feedback to be located in the cursors KCM, ignoring
the launch feedback KCM located below it.

This moves its configuration to be consistent with user expectations.